aws/webapp/requirements.txt
flask
boto3
pillow
